+++ examples/sdmmc.rs
@@ -145,15 +145,13 @@ fn main() -> ! {
     info!("");
 
     // Read test
-    let mut buffer = [0u8; 5120];
+    let mut buffer = [0u8; 512 * 10];
 
     cp.DWT.enable_cycle_counter();
     let start = pac::DWT::cycle_count();
 
-    for i in 0..10 {
-        // Read 10 blocks
-        sdmmc.read_blocks(10 * i, &mut buffer).unwrap();
-    }
+    // Read 10 blocks
+    sdmmc.read_blocks(0, &mut buffer).unwrap();
 
     let end = pac::DWT::cycle_count();
     let duration = (end - start) as f32 / ccdr.clocks.c_ck().raw() as f32;