+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Update `smoltcp` dependency to `0.9.0`
 * MSRV increased to 1.65.0
 * add `IntoAf` trait to restrict `into_alternate` [#346]
+* sdmmc: Fix read speed test.
 
 ## [v0.14.0] 2023-03-22